feat(repeat): add higher-order lettable version of repeat
