+++ src/operators/index.ts
@@ -53,6 +53,7 @@ export { publish } from './publish';
 export { publishBehavior } from './publishBehavior';
 export { race } from './race';
 export { reduce } from './reduce';
+export { repeat } from './repeat';
 export { refCount } from './refCount';
 export { scan } from './scan';
 export { subscribeOn } from './subscribeOn';
